Add files via upload
will help the registrar’s office implement a program, which will analyze the grade data that have been recorded and generate an appropriate report. 

Some the enrolled students have not yet paid their tuition yet. If a student has paid the tuition, the students’ grades are shown on grade report with the grade-point average (GPA). If not, the students’ grades are not displayed. For those students who have not paid the tuition, the grade report shall contain an alert stating that the grades have been held for nonpayment of the tuition. The grade report also shows the billing amount.